+++ Sources/BraintreeCard/BTCardClient_Internal.h
@@ -18,11 +18,6 @@ NS_ASSUME_NONNULL_BEGIN
 */
 + (NSDictionary *)validationErrorUserInfo:(NSDictionary *)userInfo;
 
-/**
- The `PPDataCollector` class, exposed internally for injecting test doubles for unit tests
- */
-+ (void)setPayPalDataCollectorClass:(nonnull Class)payPalDataCollectorClass;
-
 @end
 
 NS_ASSUME_NONNULL_END