+++ CHANGELOG.md
@@ -6,8 +6,11 @@
   * Remove `BTTokenizationService`
   * Make `BTPaymentMethodNonceParser` private
   * Remove `initWithNumber` and `initWithParameters` initializers from `BTCard`
+  * Replace `BTVenmoDriver.authorizeAccount` methods with `BTVenmoDriver.tokenizeVenmoAccount`
 * Fix memory leak in `BTPayPalDriver`
 * Add `offerPayLater` to `BTPayPalRequest`
+* Add `environment` to `BTConfiguration`
+* Add `BTVenmoRequest`
 
 ## 5.0.0-beta2 (2021-01-20)
 * Add SPM support for `BraintreeDataCollector` and `BraintreeThreeDSecure`